Added logic for conditional rendering for all pages in App.jsx by creating a top level context to set what page should be viewable.
This includes some added components for pages that are currently empty, and those file names may be changed with time. They are placeholders for now.

Working functionality right now includes being able to navigate from the "Fresh Talent" section to the artist profile page, and to navigate back to the home page by clicking "Buskin" or the logo in the nav bar.